refactor: migrate AuthTypes
