--- src/auth/dto/google-profile.ts
@@ -1,6 +1,0 @@
-export type GoogleProfile = {
-  id: string;
-  email: string;
-  verified_email: boolean;
-  picture: string;
-};
