--- src/auth/dto/index.ts
@@ -1,2 +1,0 @@
-export { GoogleProfile } from './google-profile';
-export { JwtPayload } from './jwt-payload';
