--- src/auth/dto/jwt-payload.ts
@@ -1,3 +1,0 @@
-export type JwtPayload = {
-  id: string;
-};
